Integrate test_readme.py to other files
